chore(admin-cli): rename target-templates to config-templates
