+++ admin-cli/config-templates/templates.js
